requirements.txt
tensorflow_datasets
pycocotools
tensorflow_addons